Add files via upload
By using Android technology we are trying to solve accident detection and providing emergency services to the victim like Police, Ambulance,Hospital and report to the victim's relative. To detect accident we are using Accelerometer sensor and trilateral algorithm for getting exact position ie. GPS, and GSM system for reporting to the victims relative. To implement this project i'm using Android Studio for making android application, Eclipse for provide and manage web services, MySql for backend database.